requirement.txt
Django==1.6.1
argparse==1.2.1
coverage==3.7.1
django-tastypie==0.11.0
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.5.2
wsgiref==0.1.2
